Created API docs for the surface plot classes.
Conflicts:

	surfaceplot/SurfacePlot.java
	surfaceplot/SurfacePlotGroup.java

+++ ij3d/Image3DUniverse.java
@@ -412,7 +412,7 @@ public class Image3DUniverse extends DefaultAnimatableUniverse {
 		scene.addChild(c);
 		contents.put(c.name, c);
 		recalculateGlobalMinMax(c);
-		float range = globalMax.x - globalMin.x;
+		float range = (float)(globalMax.x - globalMin.x);
 		ensureScale(range);
 		fireContentAdded(c);
 		return c;
